==> 0-runtest.ipynb <==
import os

def list_files(startpath):
    for root, dirs, files in os.walk(startpath):
        level = root.replace(startpath, '').count(os.sep)
        indent = ' ' * 4 * (level)
        print('{}{}/'.format(indent, os.path.basename(root)))
        subindent = ' ' * 4 * (level + 1)
        for f in files:
            print('{}{}'.format(subindent, f))
list_files("../")

%run 1-0-data_preprocess.ipynb
%run 1-1-EdgePrepare.ipynb
%run 1-2-Graph_embeddings.ipynb
%run 2-1-FeatureExact.ipynb
%run 2-2-extract_features.ipynb
%run 3-1-lgb_embedding.ipynb
%run 3-2-lightgbm_model.ipynb
%run 4-1-blend_model.ipynb
%run 4-2-rule_adjust.ipynb

# import pandas as pd
df = pd.read_csv("../cache/2018-12-16_737044/device1__embtype_emb_test_0_.csv")
df1 = pd.read_csv("../cache/2018-12-16_737044/device_code2__embtype_emb_test_1_.csv")

import numpy as np

a = np.zeros(df.shape[0])
b = np.zeros(df.shape[0])

a[df.index] = df.Tag.values
b[df1.index] = df1.Tag.values

df.Tag.values,df1.Tag.values

np.concatenate( [df.Tag.values.reshape(-1,1),df1.Tag.values.reshape(-1,1)  ] ,axis=1) 

df.Tag.values.reshape(-1,1)

y_train.ravel()

len(y_train)
